Rename current_keymap.* to keymaps.*
