+++ right/src/usb_report_updater.c
@@ -3,7 +3,7 @@
 #include "led_display.h"
 #include "layer.h"
 #include "usb_interfaces/usb_interface_mouse.h"
-#include "current_keymap.h"
+#include "keymaps.h"
 #include "test_states.h"
 #include "peripherals/test_led.h"
 #include "slave_drivers/is31fl3731_driver.h"
